requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.1.8
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
et_xmlfile==2.0.0
flake8==7.2.0
iniconfig==2.1.0
isort==6.0.1
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==2.2.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
pycodestyle==2.13.0
pyflakes==3.3.2
pytest==8.3.5
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
